+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+### 12/10/13 - Commit 1 ###
+* As per Issue #5 you can now edit jobs
+* Removing jobs are about 1/2 done
+* Also made the removing employees view including pagination
+
 ### 12/9/13 - Commit 1 ###
 * https://github.com/Ryuske/TimeClock/issues/5 <-- Started integrating issue #5
-** Can add/update/remove clients
-** Can add jobs
+* Can add/update/remove clients
+* Can add jobs
 * Updated main controller comments
 
 ### 12/6/13 - Commit 2 ###